tests: add test for exports
